Fixes #35702 - Move Card Context to apply all host tabs
