+++ webpack/assets/javascripts/react_app/components/HostDetails/Tabs/Details/index.js
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import React, { useEffect, useReducer, useCallback } from 'react';
+import React, { useEffect, useContext } from 'react';
 import { Flex, FlexItem, Button } from '@patternfly/react-core';
 import { registerCoreCards } from './CardRegistry';
 import Slot from '../../../common/Slot';
@@ -7,48 +7,7 @@ import { STATUS } from '../../../../constants';
 import '../Overview/styles.css';
 import './styles.css';
 import { translate as __ } from '../../../../common/I18n';
-
-export const CardExpansionContext = React.createContext({});
-
-const cardExpansionReducer = (state, action) => {
-  // A React reducer, not a Redux one!
-  switch (action.type) {
-    case 'expand':
-      return {
-        ...state,
-        [action.key]: true,
-      };
-    case 'collapse':
-      return {
-        ...state,
-        [action.key]: false,
-      };
-    case 'add':
-      if (state[action.key] === undefined) {
-        return {
-          ...state,
-          [action.key]: true,
-        };
-      }
-      return state;
-    case 'expandAll': {
-      const expandedState = {};
-      Object.keys(state).forEach(key => {
-        expandedState[key] = true;
-      });
-      return expandedState;
-    }
-    case 'collapseAll': {
-      const collapsedState = {};
-      Object.keys(state).forEach(key => {
-        collapsedState[key] = false;
-      });
-      return collapsedState;
-    }
-    default:
-      throw new Error(`invalid card expansion type: ${action.type}`);
-  }
-};
+import { CardExpansionContext } from '../../CardExpansionContext';
 
 const DetailsTab = ({ response, status, hostName }) => {
   useEffect(() => {
@@ -58,8 +17,7 @@ const DetailsTab = ({ response, status, hostName }) => {
     registerCoreCards();
     return () => document.body.classList.remove('pf-gray-background');
   }, []);
-
-  const [cardExpandStates, dispatch] = useReducer(cardExpansionReducer, {});
+  const { cardExpandStates, dispatch } = useContext(CardExpansionContext);
   const areAllCardsExpanded = Object.values(cardExpandStates).every(
     value => value === true
   );
@@ -67,25 +25,6 @@ const DetailsTab = ({ response, status, hostName }) => {
   const expandAllCards = () => dispatch({ type: 'expandAll' });
 
   const collapseAllCards = () => dispatch({ type: 'collapseAll' });
-
-  // On mount, get values from localStorage and set them in state
-  const initializeCardFromLocalStorage = useCallback(key => {
-    const value = localStorage.getItem(`${key} card expanded`);
-    if (value !== null && value !== undefined) {
-      dispatch({ type: value === 'true' ? 'expand' : 'collapse', key });
-    } else {
-      dispatch({ type: 'add', key });
-    }
-  }, []);
-
-  // On unmount, save the values to local storage
-  // eslint-disable-next-line arrow-body-style
-  useEffect(() => {
-    return () =>
-      Object.entries(cardExpandStates).forEach(([key, value]) =>
-        localStorage.setItem(`${key} card expanded`, value)
-      );
-  });
 
   const buttonText = areAllCardsExpanded
     ? __('Collapse all cards')
@@ -109,21 +48,13 @@ const DetailsTab = ({ response, status, hostName }) => {
         flexWrap={{ default: 'wrap' }}
         className="details-tab-flex-container"
       >
-        <CardExpansionContext.Provider
-          value={{
-            cardExpandStates,
-            dispatch,
-            registerCard: initializeCardFromLocalStorage,
-          }}
-        >
-          <Slot
-            hostDetails={response}
-            status={status}
-            hostName={hostName}
-            id="host-tab-details-cards"
-            multi
-          />
-        </CardExpansionContext.Provider>
+        <Slot
+          hostDetails={response}
+          status={status}
+          hostName={hostName}
+          id="host-tab-details-cards"
+          multi
+        />
       </Flex>
     </div>
   );
