+++ webpack/assets/javascripts/react_app/components/HostDetails/Templates/CardItem/CardTemplate/index.js
@@ -13,7 +13,7 @@ import {
   FlexItem,
 } from '@patternfly/react-core';
 
-import { CardExpansionContext } from '../../../Tabs/Details';
+import { CardExpansionContext } from '../../../CardExpansionContext';
 
 const CardTemplate = ({
   header,
